Analysis d merged (#58)
* merged with dev

* added files

* path bug, but word starts now

* wohoo, creates report

* bookmarks are split to reportgen by category

* tables can now be created, working on image formatting

* create report with table now works

* tables in tables

* one cat works

* 2x2table in row 2

* working report!

* working for large test report

* cleanned up and fixed category name

* removed out commented threadcode

* commented and done

* new paragraph in image text => new line

* compile for my system

* organized map structure and added videostate

* video remembers its frame when switching between videos

* renamed files

* working on analysis settings

* working settings (atleast sample freq), now just for the widget

* rectangle choice for analysis

* removed debug messeges

* dialog

* added dialog file

* owrking on advanced analysis dialog

* settings and no controll

* analysis crash

* working advanced analysis widget, using threadpool for analyses since it improved performance for switching between analyses

* fixed error with saving recent project, moved some code around

* removed debugs

* removed debugs

* mend

* working on 4k issue

* analysis queue now has widget where order can be rearranged

* fixed case for empty queue

* works with less matrices

* removed files

* set sample rate correctly

* ready for action

* removed unnecesary clone

* removed debugs

* removed 2 debugs

* remved empty lines

* deleted unused files

+++ ViAn/reportgenerator.cpp
@@ -108,7 +108,7 @@ void ReportGenerator::create_bookmark_table(QAxObject* para, ReportContainer rp_
 
     //Table indexed from 1, begin after title.
     int cell_row = 2;
-    for (int i = 0; i != rp_cont.size(); i++) { // for each category, make a paragraph of bookmarks
+    for (size_t i = 0; i != rp_cont.size(); i++) { // for each category, make a paragraph of bookmarks
         //Acess duplicate category title cells and merge them together
         QAxObject* _tmp_title = table->querySubObject("Cell(int,int)", cell_row, 1);
         QAxObject* _tmp_title2 = table->querySubObject("Cell(int,int)", cell_row, 2);
